refactor(lua): experiment with other param/result passing modes
As noted, we want to minimize the amount of memory we pass around. I
think what I am going to do is have a scanner hold the Lua strings, try
to pass a reference over to C (without copying), and then return an
array of matching indices; we'll use GC on the scanner to tell it to
destruct and free any internal state it has (I expect it will want to
cache the candidates array just like the Ruby implementation currently
does in an instance variable).