fix: make match.c compile without Ruby
Note some TODOs in here. We don't want to call `strlen()` (expensive),
so a future refactor will look at storing strings in Ruby-like structs
that include both a pointer to the storage as well as a length field.
The changes here are just minimal to make the thing compile without
warnings.