feat(lua): use Makefile trick to rebuild when changing DEBUG value
Based on trick here:

    https://stackoverflow.com/a/26147844/2103996

A half-dozen lines that dumps `CCFLAGS` contents into `.make/CCFLAGS`,
and will rebuild if they change.

I chose this method because it is simple enough to understand, and much,
much more minimal than alternatives like this one:

    https://www.cmcrossroads.com/article/rebuilding-when-cppflags-changes

and more static than this one:

    http://www.bnikolic.co.uk/blog/sh/make/unix/2021/07/08/makefile.html

The down side? It no longer prints:

    make: Nothing to be done for `all'.

when there is nothing to do (because it _always_ runs the phony target).

Note, I could have done this slightly less magically as follows:

  .PHONY: variable

  .make/CCFLAGS: variable
          @if [[ `cat .make/CCFLAGS 2>&1` != '$(CCFLAGS)' ]]; then echo $(CCFLAGS) > .make/$CCFLAGS; fi