fix(lua): don't leak transformed needles
I kind of hate this (actually, drop the "kind of"; it's cleaner), but I
am very tired right now and I more elegant/clear pattern doesn't occur
to me.

As the comments try to indicate, this works because:

- We don't "own" the passed in `needle`; we don't need it to survive
  long, because we pass it to `get_matches()` and it doesn't have to
  live beyond that.
- BUT, we do keep a copy in the form of `last_needle` as an
  optimization.
- The presence of `downcased_needle` and `squished_needle` complicates
  things, because those may be transformed versions of the original
  passed in `needle`.
- All that means that the `last_needle` we keep may be a copy of
  `needle`, or a copy of `downcased_needle` or `squished_needle`.
- Either way, we free `last_needle` in the end.

I could avoid the copy of `downcased_needle` or `squished_needle`, but I
didn't want to complicate things even further in this first pass. I will
sleep on it and see if I come up with anything better.

The idea I am playing with is this: we have to make a copy anyway, so we
may as well make the copy and then mutate it in-place. Then we free the
copy. I think that will be less ugly.