fix(lua): fix scoring bug
Our check here wasn't good enough. It is possible for `needle_idx` to be
`0` and still not have actually found the needle during the prescan (ie.
still be looking for the last/first character of the needle). So, add in
an ugly `found_needle` (`bool`) to track it.

Still some sorting bug to deal with, because two runs in a row produce
different results (searching for "tagsf"):

    {"PHP_vintage_case_default_indent", "tags-file-format", "g:rust_use_custom_ctags_defs", "tag-skip-file"}
    {"tags-file-format", "g:rust_use_custom_ctags_defs", "PHP_vintage_case_default_indent", "tag-skip-file"}

But logging shows the _scores_ are definitely stable, as is the apparent
order in which things are logged (the order may not be stable though,
due to vagaries of threading):

-   First run:

    score 0.380322 mask 421e507d for g:rust_use_custom_ctags_defs
    score 0.279539 mask 403ca9fd for PHP_vintage_case_default_indent
    score 0.630000 mask e5971 for tags-file-format
    score 0.636923 mask c9d71 for tag-skip-file

-   Second run:

    score 0.380322 mask 421e507d for g:rust_use_custom_ctags_defs
    score 0.279539 mask 403ca9fd for PHP_vintage_case_default_indent
    score 0.630000 mask e5971 for tags-file-format
    score 0.636923 mask c9d71 for tag-skip-file