test(lua): finish porting existing matcher tests
This gives us confidence that the ported library matches the old one,
fix for fix, and bug for bug (see the one test marked with `pending`).

For context on that bug, see 41a95ee4, ("Add (pending) failing test for
dotfile edge case", February 28, 2016):

> Add (pending) failing test for dotfile edge case
>
> This bug exists in 3.0.2 and I believe for quite a while before that.
> It's possible it was always present. The issue is that we stop looking
> as soon as we find a match, and that means we might not notice that
> something is a dotfile.
>
> The solution will be to check for dotfiles during our right-to-left
> prescan. Shouldn't cost us any time, and might even enable us to
> simplify the implementation of `recursive_match`. Not going to fix it
> right now because I am in the middle of this heap experiment; will
> come back to it.

6.5 years old, and still going strong as ever! But the commit message
gives me hope that we can expect to fix this with no downside.