perf(lua): add `find` scanner to scanner benchmarks
Added some `collectgarbage()` calls in there to ensure that we don't run
out of memory when running these scanners hundreds of times in a row. I
get the impression that the Lua collector is either pretty laid back
about when to run, or I am leaking like crazy (probably the latter).