refactor(lua): get rid of `g:CommandTSuppressRubyDeprecationWarning`
As noted in the previous commit, this seems like the right thing to do,
allowing us to get rid of a potentially confusing and redundant setting.

The only downside here is that we lose the prompt to automatically apply
the setting. Strictly speaking, I didn't _have_ to remove that, but it
felt a bit intrusive, and the `input()` call _did_ mess up the colors
like I said. So, this means the user _has_ to take action in order to
avoid seeing the message every time. That might be a bit annoying,
forcing a choice on the user, but it will at least ensure awareness.
Fingers crossed, I won't come to regret this decision.

Oh well, a new prompt like this _is_ a breaking change no matter what,
which is why this is going out with a major version bump (although as
noted elsewhere, SemVer doesn't mean much in the Vim world).

See: https://github.com/wincent/command-t/issues/391