docs: update HISTORY section with notes from 5.0.5 release
This is a cherry pick of 65646d2e2d13246f978f455d7422a645e95bc4f5
("docs: update HISTORY section prior to release", Jul 25 2022), on the
"5-x-release" branch. That may well be the last release in the 5.x
series for a long, long time. "pu" will go to "next"/"main" soon, and
we'll call it 6.0.0-a.0 (or maybe 6.0.0-a.1; I'm not sure yet).