feat(lua): teach Watchman layer to degrade gracefully
Lots of nasty error handling in here.

Prior to this commit, we would just `abort()` as soon as something went
wrong, but it is too easy for things to go wrong in routine operation,
so we had best handle the errors.

This would be easy in a language with exceptions. In C, I can either use
`setjmp()`/`longjmp()` to implement something exception-like, or I can
thread through explicit error state everywhere. I did the latter, even
though it is fiddly. `setjmp()`/`longjmp()` sounds expensive, and it
robs you of any chance to do intermediate cleanup as you unwind the
stack.

I don't think I need to do any intermediate clean-up in this specific
instance, but it's nice to have that capability if needed.

So with this change, we catch errors at the Lua layer and surface them
with `error()`. This may seem mean, but the old version aborted the
entire process. As noted in the comments, a final version of this would
just quietly log the error and fall back to an alternative scanner
instead, but for now, I want it to fail loudly so I can fix it when
things go wrong.