fix(lua): restore current window before running open command
Remember the current window before going to the UI, then restore it when
coming back.

Otherwise, we end up always opening in the topmost/leftmost split,
because that is where Neovim is going to put the focus when we close.
Likewise if we dismiss the UI without choosing an item, we would go back
to the topmost/leftmost split.

Not sure if I need to defer until after the autocommands run or not, but
it doesn't seem like it hurts to do so, so using `defer_fn()` for that.

Still not solved: my custom autocommands don't cope well with the UI
being dismissed without actually opening an item. Focus goes back to the
window you came from, but the syntax etc doesn't get reapplied. Will dig
into that later.