refactor(lua): remove autocmd that does nothing
Not sure why, but I never see any autocmds run for these title buffers
(I tried adding a bunch and _none_ of them get called). Additionally,
`buftype` seems to be doing nothing. So, removing the code that doesn't
seem to be doing what I think it should be doing, at least for now.