feat(lua): set buffer name, for debuggability
To assist me with a mysterious additional scratch buffer that I am not
expecting.

When I open Neovim with no file, `:ls!` shows this:

    1 %a   "[No Name]"                    line 1

After opening Command-T, I see this (edited for readability; this is
after adding a `vim.cmd('ls!')` at the end of `Window:show()`:

    1 #a   "[No Name]"                    line 1
    2u%a   "CommandT [match listing] (main)" line 1

This is after the match listing is shown. We have the main window, and
no title. That's at the end of the first call to `Window:show()`.

    1  a   "[No Name]"                    line 1
    2u a   "CommandT [match listing] (main)" line 1

Then we create the prompt, so we see the same entries repeated again,
then...

    3u%a + "CommandT [file] (title)"      line 1
    4u a   "[Scratch]"                    line 0
    5u#    "CommandT [file] (main)"       line 1

Buffers 3 and 5 are expected; as their names indicated, they are the
main and title buffers for the prompt area. But what is that mysterious
buffer number 4?

Additionally, once I close Command-T, buffer 5 lingers around; `:ls!`
shows:

    1 %a   "[No Name]"                    line 1
    5u     "CommandT [file] (main)"       line 1