refactor: get rid of diagnostic message
This is in response to this feedback:

    https://github.com/wincent/command-t/issues/393#issuecomment-1229233889

The user in question is loading Command-T like this, using packer:

    use {
      'wincent/command-t',
      run = 'cd lua/wincent/commandt/lib && make',
      config = function()
        vim.g.CommandTPreferredImplementation = 'lua'
        require('wincent.commandt').setup({})
      end
    }

and they see this whenever they run `:PackerCompile`:

    commandt.setup():
      `commandt.setup()` was called after Ruby plugin setup has already run

So, I'm removing this check as it doesn't add a lot of value. My advice,
then, is to move this statement:

    vim.g.CommandTPreferredImplementation = 'lua'

up into the top-level of the `~/.config/nvim/init.lua`. That way,
whenever Command-T's `plugin/command-t.vim` is sourced, it will do the
right thing, which means:

1. Skip doing anything if `command_t_loaded` is already set.
2. Do Lua-centric set-up instead of Ruby-centric set-up.

And then when `plugin/command-t.lua` is sourced (which will always
happen second, because Neovim loads Lua _after_ Vimscript), it will
again do the right thing (ie. do the Lua-centric instead of Ruby-centric
set-up).